requirements.txt
autopep8==2.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
pycodestyle==2.10.0
pytest==7.3.2
pytest-mock==3.10.0
requests==2.31.0
robotframework==6.1
urllib3==2.0.3
